requirements.txt
openai-whisper
gtts
spacy
streamlit
groq
pytest
ffmpeg-python